// frontend/src/shared/types.ts
export type Plan = 'free' | 'pro'
export type UserStatus = 'trial' | 'pending' | 'active' | 'suspended'
export type AccountType = 'demo' | 'real'
// ...
export type IndexSymbol =
  | 'CRASH_300' | 'CRASH_500' | 'CRASH_600' | 'CRASH_900' | 'CRASH_1000'
  | 'BOOM_300'  | 'BOOM_500'  | 'BOOM_600'  | 'BOOM_900'  | 'BOOM_1000'
// ...
export interface User {
  userId: string
  email: string
  name: string
  plan: Plan
  status: UserStatus
  trialEndsAt: string
  subscriptionPrice: number
  derivToken?: string
  derivAccountType?: AccountType
  derivLoginId?: string
  metaApiAccountId?: string
  mt5LoginId?: string
  mt5Server?: string
  mt5AccountType?: AccountType
  mt5LastSync?: string
  subscriptionStart?: string
  subscriptionEnd?: string
  createdAt: string
  watchlist: IndexSymbol[]
  pushSubscription?: PushSubscriptionData
}

export interface PushSubscriptionData {
  endpoint: string
  keys: { p256dh: string; auth: string }
}